+++ src/app/components/forms/tag-select/example/terra-tag-select.component.example.ts
@@ -11,11 +11,11 @@ import { Component } from '@angular/core';
 export class TerraTagSelectComponentExample
 {
     protected selectedTags:Array<TerraTagInterface> = [{
-        id:    3,
-        name:  'Fast schwarz',
-        color: '#000350',
+        id:         3,
+        name:       'Fast schwarz',
+        color:      '#000350',
         isClosable: true,
-        names: [
+        names:      [
             {
                 id:       3,
                 tagId:    3,