+++ src/app/components/tables/data-table/context-menu/terra-data-table-context-menu.service.ts
@@ -12,5 +12,5 @@ export class TerraDataTableContextMenuService<D extends TerraBaseData>
      * @description channel for component <-> directive communication which is used to open the context menu on click
      */
     public show:Subject<{ event:MouseEvent, data:D }> =
-        new Subject<{ event:MouseEvent, data:D}>();
+        new Subject<{ event:MouseEvent, data:D }>();
 }