+++ src/polyfills.ts
@@ -1,6 +1,7 @@
 import 'core-js/es6';
 import 'core-js/es7/reflect';
-import 'zone.js/dist/zone';
+
+require('zone.js/dist/zone');
 
 if(process.env.ENV === 'production')
 {
@@ -8,8 +9,7 @@ if(process.env.ENV === 'production')
 }
 else
 {
-    // Development
+    // Development and test
     Error['stackTraceLimit'] = Infinity;
-
     require('zone.js/dist/long-stack-trace-zone');
 }