+++ src/app/components/forms/form/model/form-type-map.ts
@@ -1,27 +1,21 @@
-import {
-    TerraCategoryPickerComponent,
-    TerraCheckboxComponent,
-    TerraCodeEditorComponent,
-    TerraColorPickerComponent,
-    TerraDatePickerComponent,
-    TerraDoubleInputComponent,
-    TerraFileInputComponent,
-    TerraMultiCheckBoxComponent,
-    TerraNoteEditorComponent,
-    TerraNumberInputComponent,
-    TerraSelectBoxComponent,
-    TerraSliderComponent,
-    TerraTextAreaInputComponent,
-    TerraTextInputComponent
-} from '../../../../../';
 import { Type } from '@angular/core';
+import { CheckboxGroupComponent } from '../../checkbox-group/checkbox-group.component';
 import { TerraFormTypeInterface } from './terra-form-type.interface';
+import { TerraCheckboxComponent } from '../../checkbox/terra-checkbox.component';
+import { TerraDatePickerComponent } from '../../input/date-picker/terra-date-picker.component';
+import { TerraFileInputComponent } from '../../input/file-input/terra-file-input.component';
+import { TerraTextInputComponent } from '../../input/text-input/terra-text-input.component';
+import { TerraTextAreaInputComponent } from '../../input/text-area-input/terra-text-area-input.component';
+import { TerraNumberInputComponent } from '../../input/number-input/terra-number-input.component';
+import { TerraDoubleInputComponent } from '../../input/double-input/terra-double-input.component';
+import { TerraSelectBoxComponent } from '../../select-box/terra-select-box.component';
+import { TerraCategoryPickerComponent } from '../../../data-picker/category-picker/terra-category-picker.component';
+import { TerraColorPickerComponent } from '../../input/color-picker/terra-color-picker.component';
+import { TerraNoteEditorComponent } from '../../../editors/note-editor/terra-note-editor.component';
+import { TerraCodeEditorComponent } from '../../../editors/code-editor/terra-code-editor.component';
+import { TerraSliderComponent } from '../../slider/terra-slider.component';
 
-/**
- * @deprecated use FormTypeMap instead. It now uses the `CheckboxGroupComponent` to render a 'checkboxGroup'.
- * TODO: If any changes to this class are necessary, please apply the changes to the FormTypeMap as well.
- */
-export class TerraFormTypeMap
+export class FormTypeMap
 {
     public readonly checkbox:TerraFormTypeInterface = {
         inputMap:  {
@@ -78,7 +72,7 @@ export class TerraFormTypeMap
         },
         component: TerraColorPickerComponent
     };
-    public readonly checkboxGroup:Type<TerraMultiCheckBoxComponent> = TerraMultiCheckBoxComponent;
+    public readonly checkboxGroup:Type<CheckboxGroupComponent> = CheckboxGroupComponent;
     public readonly slider:Type<TerraSliderComponent> = TerraSliderComponent;
     public readonly noteEditor:TerraFormTypeInterface = {
         inputMap:  {
@@ -92,9 +86,4 @@ export class TerraFormTypeMap
         },
         component: TerraCodeEditorComponent
     };
-
-    constructor()
-    {
-        console.warn('Usage of deprecated class `TerraFormTypeMap`. Please consider using `FormTypeMap` instead.');
-    }
 }