+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Bug Fixes
 
-* **terra-category-picker** REST calls will now be triggered once.
+* **terra-category-picker** REST calls will no longer be triggered twice.
 
 <a name="3.2.0"></a>
 # 3.2.0 (23.01.2019)