+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * **webpack update** updated to latest version of webpack due to security vulnerabilities
 
 ### Bug Fixes
-* **terra-form** fixed initialization of form elements with defaultValues
+* **terra-form-entry** fixed initialization with defaultValue
 
 <a name="3.2.0"></a>
 # 3.2.0 (23.01.2019)