+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<a name="3.X.X"></a>
+# 3.X.X (XX.XX.2019)
+
+### Features 
+* **webpack update** updated to latest version of webpack due to security vulnerabilities
+
 <a name="3.2.0"></a>
 # 3.2.0 (23.01.2019)