+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 
 ### Features 
 * **webpack update** updated to latest version of webpack due to security vulnerabilities
-* **terra-slider** added to **terra-form**
+* **terra-form** added support for ´<terra-slider>`
+
+### Bug Fixes
+* **terra-form-entry** fixed initialization with defaultValue
+
 <a name="3.2.0"></a>
 # 3.2.0 (23.01.2019)