+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Bug Fixes
 * **terra-form-entry** fixed initialization with defaultValue
+* **terra-form-entry** check if component property is decorated before assigning value dynamically
 
 <a name="3.2.0"></a>
 # 3.2.0 (23.01.2019)